implemented safer deletion call

+++ website/api/api-post.php
@@ -115,16 +115,19 @@ if ($_GET["action"] == "fetch") {
     $response["postDeleted"] = false;
     $postId = $_POST["postId"];
 
-    $post = getPostFromPostId($postId, $mysqli);
-    $username = getUser($_SESSION["user_id"], $mysqli)['username'];
+    if (userCanDelete($postId, $_SESSION['user_id'], $mysqli)) {
+      $post = getPostFromPostId($postId, $mysqli);
+      $username = getUser($_SESSION["user_id"], $mysqli)['username'];
 
-    list($res, $msg) = deleteImg(POST_IMG_DIR. $username . "/posts/". $post["image"]);
+      list($res, $msg) = deleteImg(POST_IMG_DIR. $username . "/posts/". $post["image"]);
 
-    $response["msg"] = $msg;
-    
-    if ($res && deletePost($postId, $mysqli)) {
-      $response["postDeleted"] = true;
+      $response["msg"] = $msg;
+      
+      if ($res && deletePost($postId, $mysqli)) {
+        $response["postDeleted"] = true;
+      }
     }
+
     header("Content-Type: application/json");
     echo json_encode($response);
     return;
